Remove mapper and clean up code
